Increase code coverage for subscription logic
